// mh-personal-trainer-react-native/src/types/feedback.ts
export interface FeedbackItem {
  id: string;
  uid?: string;
  codigoDoPersonal?: number;
  yourName?: string;
  comentarioDoAluno?: string;
  respostaDoProf?: string;
  estrela?: number;
  horaDeTermino?: Date;
  melhoria?: string;
  dificuldadeDoExercicio?: string;
  progresso?: string;
  tempoDoTreino?: number;
  imgUser?: string;
  nomeDoTreino?: string;
  obsInstrucao?: string;
  refPath?: string;
}